requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
chromadb==1.0.8
fastapi==0.115.12
jose==1.0.0
langchain==0.3.25
langchain_community==0.3.23
langchain_core==0.3.58
langchain_experimental==0.3.4
langchain_ollama==0.3.2
langgraph==0.4.2
langgraph_supervisor==0.0.20
matplotlib==3.10.1
passlib==1.7.4
pydantic==2.11.4
python-dotenv==1.1.0
python_jose==3.4.0
SQLAlchemy==2.0.37
typing_extensions==4.13.2
uvicorn==0.34.2
python-multipart==0.0.20
langgraph-checkpoint-sqlite==2.0.7

networkx==3.4.2
unstructured==0.17.2
